Fix worker logs
