+++ locust/webui/src/components/LogViewer/LogViewer.utils.tsx
